+++ packages/camera/android/src/test/java/io/flutter/plugins/camera/CameraPermissionsTest.java
@@ -1,0 +1,23 @@
+package io.flutter.plugins.camera;
+
+import static junit.framework.TestCase.assertEquals;
+
+import android.content.pm.PackageManager;
+import io.flutter.plugins.camera.CameraPermissions.CameraRequestPermissionsListener;
+import org.junit.Test;
+
+public class CameraPermissionsTest {
+  @Test
+  public void listener_respondsOnce() {
+    final int[] calledCounter = {0};
+    CameraRequestPermissionsListener permissionsListener =
+        new CameraRequestPermissionsListener((String code, String desc) -> calledCounter[0]++);
+
+    permissionsListener.onRequestPermissionsResult(
+        9796, null, new int[] {PackageManager.PERMISSION_DENIED});
+    permissionsListener.onRequestPermissionsResult(
+        9796, null, new int[] {PackageManager.PERMISSION_GRANTED});
+
+    assertEquals(1, calledCounter[0]);
+  }
+}